Rename math.py → mathbrain.py and update README
- Fix module name conflict with Python builtin math module
- Add mathbrain section to README (philosophy + Phase 1/2 explanation)
- Update test count: 135 → 166 tests
- Add test_math.py to test list
- All 31 MathBrain tests now passing (0.249s)

Philosophy: MathBrain is leo's body awareness — proprioception through mathematics.
